move function to base.py
